rename drawer component to NavDrawer
