+++ src/components/Home/Home.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Text, View, StyleSheet } from 'react-native';
-import Drawer from '../Drawer/Drawer';
+import Drawer from '../NavDrawer/ NavDrawer';
 import Welcome from '../Welcome/Welcome';
 export default class Home extends Component {
 	render () {
